requirements.txt
python-dotenv
requests
openai
gradio
ollama
